address Win32 analysis warnings

+++ tests/services/AchievementRuntime_Tests.cpp
@@ -462,7 +462,7 @@ public:
         pAchievement3->SetConditionHitCount(0, 0, 1);
         pAchievement5->SetConditionHitCount(0, 0, 1);
         nSize = runtime.SaveProgressToBuffer(sBuffer.data(), nSize);
-        if (nSize > sBuffer.size())
+        if (nSize > gsl::narrow_cast<int>(sBuffer.size()))
         {
             sBuffer.resize(nSize);
             runtime.SaveProgressToBuffer(sBuffer.data(), nSize);
@@ -480,7 +480,7 @@ public:
         runtime.GetAchievementTrigger(5U)->state = RC_TRIGGER_STATE_ACTIVE;
         pAchievement5->SetConditionHitCount(0, 0, 1);
         nSize = runtime.SaveProgressToBuffer(sBuffer.data(), nSize);
-        if (nSize > sBuffer.size())
+        if (nSize > gsl::narrow_cast<int>(sBuffer.size()))
         {
             sBuffer.resize(nSize);
             runtime.SaveProgressToBuffer(sBuffer.data(), nSize);
